cleanup + prep
